@@ -10,30 +10,32 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171103073857) do
+ActiveRecord::Schema.define(version: 2019_04_04_092637) do
 
-  create_table "feeds", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+  create_table "feeds", force: :cascade do |t|
     t.string "feed_type"
     t.integer "user_id"
     t.string "image_file_name"
-    t.string "exif", :limit => 1000
+    t.string "exif", limit: 1000
     t.string "comment"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "friends", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+  create_table "friends", force: :cascade do |t|
     t.integer "from_user_id"
     t.integer "to_user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "users", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+  create_table "users", force: :cascade do |t|
     t.string "login_id"
     t.string "name"
+    t.string "email"
     t.string "pass"
     t.string "icon_file_name"
+    t.boolean "admin", default: false, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
