@@ -1,25 +1,26 @@
 'use strict';
 
 const main = () => {
-  $$('form-login').addEventListener('submit', (event) => {
+  $$('form-password-reset').addEventListener('submit', (event) => {
     event.preventDefault();
 
+    const params = getUrlParams(location.href);
+    const url = `password_reset/${params['reset_token']}`;
+
     const body = new FormData();
-    body.append('login_id', $$('input-login-id').value);
     body.append('pass', $$('input-pass').value);
 
     const process = result => {
       if (result.errors) {
-        // ログイン失敗
+        // パスワードリセット失敗
         $$('div-errors').innerHTML = '';
         result.errors.forEach(value => $$('div-errors').innerHTML += `<p class="bg-danger text-danger">${value}</p>`);
       } else {
-        // ログイン成功
+        // パスワードリセット成功
         top.location = '/'; // メイン画面を表示
       }
     }
-
-    fetcher('/sessions', { method: 'POST', body: body }, process);
+    fetcher(url, { method: 'PUT', body: body }, process);
   });
 }
 
