@@ -1,6 +1,6 @@
 'use strict';
 
-document.addEventListener('DOMContentLoaded', () => {
+const main = () => {
   $$('div-drop').addEventListener('dragover', (event) => event.preventDefault());
   $$('div-drop').addEventListener('drop', (event) => {
     event.preventDefault();
@@ -10,11 +10,11 @@ document.addEventListener('DOMContentLoaded', () => {
     $$('div-drop').textContent = '画像アップロード中';
     $$('div-drop').classList.add('drop-zone-droped');
 
-    let body = new FormData();
+    const body = new FormData();
     body.append('image', event.dataTransfer.files[0]);
     body.append('resize_max_pixel', 240);
 
-    let process = result => {
+    const process = result => {
       if (result.file_name) {
         // 画像アップロード成功
         $$('div-drop').textContent = '画像アップロード完了';
@@ -27,30 +27,31 @@ document.addEventListener('DOMContentLoaded', () => {
         $$('div-errors').innerHTML = '<p class="bg-danger text-danger">画像のアップロードに失敗しました</p>';
       }
     }
-    fetcher('/icons' ,{method: 'POST', body: body}, process);
+    fetcher('/icons', { method: 'POST', body: body }, process);
   });
 
   $$('form-register').addEventListener('submit', (event) => {
     event.preventDefault();
 
-    let body = new FormData();
-    body.append('login_id', $$('input-login-id').value);
-    body.append('name', $$('input-name').value);
-    body.append('pass', $$('input-pass').value);
-    body.append('icon_file_name', $$('input-icon-file-name').value);
+    const body = new FormData();
+    body.append('user[login_id]', $$('input-login-id').value);
+    body.append('user[name]', $$('input-name').value);
+    body.append('user[email]', $$('input-email').value);
+    body.append('user[pass]', $$('input-pass').value);
+    body.append('user[icon_file_name]', $$('input-icon-file-name').value);
 
-    let process = result => {
-      if (result.name) {
-        // ユーザ登録成功
-        localStorage.setItem('name', result.name);
-        localStorage.setItem('icon', result.icon);
-        top.location = '/'; // メイン画面を表示
-      } else {
+    const process = result => {
+      if (result.errors) {
         // ユーザ登録失敗
         $$('div-errors').innerHTML = '';
-        result.errors.forEach(value => $$('div-errors').innerHTML += `<p class="bg-danger text-danger">${value}</p>` );
+        result.errors.forEach(value => $$('div-errors').innerHTML += `<p class="bg-danger text-danger">${value}</p>`);
+      } else {
+        // ユーザ登録成功
+        top.location = '/'; // メイン画面を表示
       }
     }
-    fetcher('/users' ,{method: 'POST', body: body}, process);
+    fetcher('/users', { method: 'POST', body: body }, process);
   });
-});
+}
+
+document.addEventListener('DOMContentLoaded', main);
